# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.AppError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  profiles.CommunityProfile:
    properties:
      accepts_hot_meals:
        type: boolean
      allergens:
        items:
          type: string
        type: array
      avatar_url:
        type: string
      avoid_items:
        items:
          type: string
        type: array
      bio:
        type: string
      community_role:
        type: string
      created_at:
        type: string
      dietary_restrictions:
        items:
          type: string
        type: array
      distance_preference:
        type: string
      id:
        type: string
      notifications_enabled:
        type: boolean
      notify_on_claim:
        type: boolean
      notify_on_messages:
        type: boolean
      preferred_items:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
      username:
        type: string
      visibility:
        type: string
    type: object
  profiles.CreateProfileRequest:
    properties:
      accepts_hot_meals:
        type: boolean
      allergens:
        items:
          type: string
        type: array
      avatar_url:
        type: string
      avoid_items:
        items:
          type: string
        type: array
      bio:
        type: string
      community_role:
        type: string
      dietary_restrictions:
        items:
          type: string
        type: array
      distance_preference:
        type: string
      notifications_enabled:
        type: boolean
      notify_on_claim:
        type: boolean
      notify_on_messages:
        type: boolean
      preferred_items:
        items:
          type: string
        type: array
      username:
        maxLength: 255
        minLength: 3
        type: string
      visibility:
        type: string
    required:
    - username
    type: object
  profiles.UpdateProfileRequest:
    properties:
      accepts_hot_meals:
        type: boolean
      allergens:
        items:
          type: string
        type: array
      avatar_url:
        type: string
      avoid_items:
        items:
          type: string
        type: array
      bio:
        type: string
      community_role:
        type: string
      dietary_restrictions:
        items:
          type: string
        type: array
      distance_preference:
        type: string
      notifications_enabled:
        type: boolean
      notify_on_claim:
        type: boolean
      notify_on_messages:
        type: boolean
      preferred_items:
        items:
          type: string
        type: array
      visibility:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A Go backend API server for Foodlink application
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Foodlink Backend API
  version: "1.0"
paths:
...
  /community/profile:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's community profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profiles.CommunityProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get user's community profile
      tags:
      - community-profiles
    post:
      consumes:
      - application/json
      description: Create a new community profile for the authenticated user
      parameters:
      - description: Profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/profiles.CreateProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/profiles.CommunityProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Create community profile
      tags:
      - community-profiles
    put:
      consumes:
      - application/json
      description: Update the authenticated user's community profile
      parameters:
      - description: Profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/profiles.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profiles.CommunityProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Update community profile
      tags:
      - community-profiles
  /community/profile/{username}:
    get:
      consumes:
      - application/json
      description: Get a community profile by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profiles.CommunityProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Get profile by username
      tags:
      - community-profiles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"